cours1

+++ chap-01/4-link-errors/3-sub.cpp
@@ -1,6 +1,5 @@
 #include "3-sub.hpp"
 
-float sub(float a, float b)
-{
+int sub(int a, int b) {
     return a - b;
 }
